+++ R/02-downAndUpload.R
@@ -331,6 +331,8 @@ uploadModelUI <- function(id,
 #' @param onlySettings (logical) if TRUE allow only upload of user inputs and user data
 #' @param fileExtension (character) (optional) app specific file extension, e.g. "resources",
 #'  "bpred", "bmsc"
+#' @param extractZipFun (function) (optional) parameter to provide an app-specific function
+#'  to extract the zip object of the upload.
 #' @param mainFolder (character) folder containing all loadable .zip files. For most apps this
 #' is folder "predefinedModels". In most apps it can be found under "inst/app/".
 #' @param subFolder (character) (optional) subfolder containing loadable .zip files
@@ -348,6 +350,7 @@ uploadModelServer <-
            reloadChoices = reactive(TRUE),
            onlySettings = FALSE,
            fileExtension = "zip",
+           extractZipFun = NULL,
            reset = reactive(FALSE)) {
     moduleServer(id,
                  function(input, output, session) {
@@ -376,12 +379,14 @@ uploadModelServer <-
 
                    observeEvent(pathToModel(), {
                      withProgress({
-                       res <- loadModel(filepath = pathToModel(),
-                                        subFolder = subFolder,
-                                        rPackageName = rPackageName,
-                                        onlySettings = onlySettings,
-                                        fileExtension = fileExtension) %>%
-                         tryCatchWithWarningsAndErrors(errorTitle = "Could not load file!")
+                       res <- pathToModel() %>%
+                         loadZipWrapper(
+                           subFolder = subFolder,
+                           rPackageName = rPackageName,
+                           onlySettings = onlySettings,
+                           fileExtension = fileExtension,
+                           extractZipFun = extractZipFun
+                       )
                      },
                      value = 0.8,
                      message = "Uploading ...")